+++ enterprise/internal/rockskip/git.go
@@ -1,0 +1,358 @@
+package rockskip
+
+import (
+	"bufio"
+	"fmt"
+	"io"
+	"os/exec"
+	"strconv"
+	"strings"
+
+	"github.com/sourcegraph/sourcegraph/lib/errors"
+)
+
+type LogEntry struct {
+	Commit       string
+	PathStatuses []PathStatus
+}
+
+type PathStatus struct {
+	Path   string
+	Status StatusAMD
+}
+
+type CommitStatus struct {
+	Commit string
+	Status StatusAMD
+}
+
+type StatusAMD int
+
+const (
+	AddedAMD    StatusAMD = 0
+	ModifiedAMD StatusAMD = 1
+	DeletedAMD  StatusAMD = 2
+)
+
+type StatusAD int
+
+const (
+	AddedAD   StatusAD = 0
+	DeletedAD StatusAD = 1
+)
+
+type Git interface {
+	LogReverseEach(repo string, commit string, n int, onLogEntry func(logEntry LogEntry) error) error
+	RevListEach(repo string, commit string, onCommit func(commit string) (shouldContinue bool, err error)) error
+	ArchiveEach(repo string, commit string, paths []string, onFile func(path string, contents []byte) error) error
+}
+
+type SubprocessGit struct {
+	gitDir        string
+	catFileCmd    *exec.Cmd
+	catFileStdin  io.WriteCloser
+	catFileStdout bufio.Reader
+}
+
+func NewSubprocessGit(gitDir string) (*SubprocessGit, error) {
+	cmd := exec.Command("git", "cat-file", "--batch")
+	cmd.Dir = gitDir
+
+	stdin, err := cmd.StdinPipe()
+	if err != nil {
+		return nil, err
+	}
+
+	stdout, err := cmd.StdoutPipe()
+	if err != nil {
+		return nil, err
+	}
+
+	err = cmd.Start()
+	if err != nil {
+		return nil, err
+	}
+
+	return &SubprocessGit{
+		gitDir:        gitDir,
+		catFileCmd:    cmd,
+		catFileStdin:  stdin,
+		catFileStdout: *bufio.NewReader(stdout),
+	}, nil
+}
+
+func (git SubprocessGit) Close() error {
+	err := git.catFileStdin.Close()
+	if err != nil {
+		return err
+	}
+	return git.catFileCmd.Wait()
+}
+
+func (git SubprocessGit) LogReverseEach(repo string, givenCommit string, n int, onLogEntry func(entry LogEntry) error) (returnError error) {
+	log := exec.Command("git", LogReverseArgs(n, givenCommit)...)
+	log.Dir = git.gitDir
+	output, err := log.StdoutPipe()
+	if err != nil {
+		return err
+	}
+
+	err = log.Start()
+	if err != nil {
+		return err
+	}
+	defer func() {
+		err = log.Wait()
+		if err != nil {
+			returnError = err
+		}
+	}()
+
+	return ParseLogReverseEach(output, onLogEntry)
+}
+
+func (git SubprocessGit) RevListEach(repo string, givenCommit string, onCommit func(commit string) (shouldContinue bool, err error)) (returnError error) {
+	revList := exec.Command("git", RevListArgs(givenCommit)...)
+	revList.Dir = git.gitDir
+	output, err := revList.StdoutPipe()
+	if err != nil {
+		return err
+	}
+
+	err = revList.Start()
+	if err != nil {
+		return err
+	}
+	defer func() {
+		err = revList.Wait()
+		if err != nil {
+			returnError = err
+		}
+	}()
+
+	return RevListEach(output, onCommit)
+}
+
+func (git SubprocessGit) ArchiveEach(repo string, commit string, paths []string, onFile func(path string, contents []byte) error) error {
+	for _, path := range paths {
+		_, err := git.catFileStdin.Write([]byte(fmt.Sprintf("%s:%s\n", commit, path)))
+		if err != nil {
+			return errors.Wrap(err, "writing to cat-file stdin")
+		}
+
+		line, err := git.catFileStdout.ReadString('\n')
+		if err != nil {
+			return errors.Wrap(err, "read newline")
+		}
+		line = line[:len(line)-1] // Drop the trailing newline
+		parts := strings.Split(line, " ")
+		if len(parts) != 3 {
+			return errors.Newf("unexpected cat-file output: %q", line)
+		}
+		size, err := strconv.ParseInt(parts[2], 10, 64)
+		if err != nil {
+			return errors.Wrap(err, "parse size")
+		}
+
+		fileContents, err := io.ReadAll(io.LimitReader(&git.catFileStdout, size))
+		if err != nil {
+			return errors.Wrap(err, "read contents")
+		}
+
+		discarded, err := git.catFileStdout.Discard(1) // Discard the trailing newline
+		if err != nil {
+			return errors.Wrap(err, "discard newline")
+		}
+		if discarded != 1 {
+			return errors.Newf("expected to discard 1 byte, but discarded %d", discarded)
+		}
+
+		err = onFile(path, fileContents)
+		if err != nil {
+			return errors.Wrap(err, "onFile")
+		}
+	}
+
+	return nil
+}
+
+func LogReverseArgs(n int, givenCommit string) []string {
+	return []string{
+		"log",
+		"--pretty=%H %P",
+		"--raw",
+		"-z",
+		"-m",
+		// --no-abbrev speeds up git log a lot
+		"--no-abbrev",
+		"--no-renames",
+		"--first-parent",
+		"--reverse",
+		"--ignore-submodules",
+		fmt.Sprintf("-%d", n),
+		givenCommit,
+	}
+}
+
+func ParseLogReverseEach(stdout io.Reader, onLogEntry func(entry LogEntry) error) error {
+	reader := bufio.NewReader(stdout)
+
+	var buf []byte
+
+	for {
+		// abc... ... NULL '\n'?
+
+		// Read the commit
+		commitBytes, err := reader.Peek(40)
+		if err == io.EOF {
+			break
+		} else if err != nil {
+			return err
+		}
+		commit := string(commitBytes)
+
+		// Skip past the NULL byte
+		_, err = reader.ReadBytes(0)
+		if err != nil {
+			return err
+		}
+
+		// A '\n' indicates a list of paths and their statuses is next
+		buf, err = reader.Peek(1)
+		if err == io.EOF {
+			err = onLogEntry(LogEntry{Commit: commit, PathStatuses: []PathStatus{}})
+			if err != nil {
+				return err
+			}
+			break
+		} else if err != nil {
+			return err
+		}
+		if buf[0] == '\n' {
+			// A list of paths and their statuses is next
+
+			// Skip the '\n'
+			discarded, err := reader.Discard(1)
+			if discarded != 1 {
+				return errors.Newf("discarded %d bytes, expected 1", discarded)
+			} else if err != nil {
+				return err
+			}
+
+			pathStatuses := []PathStatus{}
+			for {
+				// :100644 100644 abc... def... M NULL file.txt NULL
+				// ^ 0                          ^ 97   ^ 99
+
+				// A ':' indicates a path and its status is next
+				buf, err = reader.Peek(1)
+				if err == io.EOF {
+					break
+				} else if err != nil {
+					return err
+				}
+				if buf[0] != ':' {
+					break
+				}
+
+				// Read the status from index 97 and skip to the path at index 99
+				buf = make([]byte, 99)
+				read, err := io.ReadFull(reader, buf)
+				if read != 99 {
+					return errors.Newf("read %d bytes, expected 99", read)
+				} else if err != nil {
+					return err
+				}
+
+				// Read the path
+				path, err := reader.ReadBytes(0)
+				if err != nil {
+					return err
+				}
+				path = path[:len(path)-1] // Drop the trailing NULL byte
+
+				// Inspect the status
+				var status StatusAMD
+				statusByte := buf[97]
+				switch statusByte {
+				case 'A':
+					status = AddedAMD
+				case 'M':
+					status = ModifiedAMD
+				case 'D':
+					status = DeletedAMD
+				case 'T':
+					// Type changed. Check if it changed from a file to a submodule or vice versa,
+					// treating submodules as empty.
+
+					isSubmodule := func(mode string) bool {
+						// Submodules are mode "160000". https://stackoverflow.com/questions/737673/how-to-read-the-mode-field-of-git-ls-trees-output#comment3519596_737877
+						return mode == "160000"
+					}
+
+					oldMode := string(buf[1:7])
+					newMode := string(buf[8:14])
+
+					if isSubmodule(oldMode) && !isSubmodule(newMode) {
+						// It changed from a submodule to a file, so consider it added.
+						status = AddedAMD
+						break
+					}
+
+					if !isSubmodule(oldMode) && isSubmodule(newMode) {
+						// It changed from a file to a submodule, so consider it deleted.
+						status = DeletedAMD
+						break
+					}
+
+					// Otherwise, it remained the same, so ignore the type change.
+					continue
+				case 'C':
+					// Copied
+					return errors.Newf("unexpected status 'C' given --no-renames was specified")
+				case 'R':
+					// Renamed
+					return errors.Newf("unexpected status 'R' given --no-renames was specified")
+				case 'X':
+					return errors.Newf("unexpected status 'X' indicates a bug in git")
+				default:
+					fmt.Printf("LogReverse commit %q path %q: unrecognized diff status %q, skipping\n", commit, path, string(statusByte))
+					continue
+				}
+
+				pathStatuses = append(pathStatuses, PathStatus{Path: string(path), Status: status})
+			}
+
+			err = onLogEntry(LogEntry{Commit: commit, PathStatuses: pathStatuses})
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	return nil
+}
+
+func RevListArgs(givenCommit string) []string {
+	return []string{"rev-list", "--first-parent", givenCommit}
+}
+
+func RevListEach(stdout io.Reader, onCommit func(commit string) (shouldContinue bool, err error)) error {
+	reader := bufio.NewReader(stdout)
+
+	for {
+		commit, err := reader.ReadString('\n')
+		if err == io.EOF {
+			break
+		} else if err != nil {
+			return err
+		}
+		commit = commit[:len(commit)-1] // Drop the trailing newline
+		shouldContinue, err := onCommit(commit)
+		if !shouldContinue {
+			return err
+		}
+	}
+
+	return nil
+}
